fix: rename migration scripts
