+++ models/migrationscripts/20220707_create_domain_tables.go
