+++ plugins/feishu/models/migrationscripts/20220714_add_init_tables.go
