+++ plugins/github/models/migrationscripts/20220715_add_init_tables.go
