+++ plugins/jira/models/migrationscripts/20220716_add_init_tables.go
